+++ common/type_system/TypeSystem.cpp
@@ -68,8 +68,32 @@ Type* TypeSystem::add_type(const std::string& name, std::unique_ptr<Type> type) 
  * This will allow the type system to generate TypeSpecs for this type, but not access detailed
  * information, or know the exact size.
  */
-void TypeSystem::forward_declare_type(std::string name) {
-  m_forward_declared_types.insert(std::move(name));
+void TypeSystem::forward_declare_type(const std::string& name) {
+  if (m_types.find(name) == m_types.end()) {
+    m_forward_declared_types[name] = TYPE;
+  }
+}
+
+/*!
+ * Inform the type system that there will eventually be a type named "name" and that it's a basic.
+ * This allows the type to be used in a few specific places. For instance a basic can have
+ * a field where an element is the same type.
+ */
+void TypeSystem::forward_declare_type_as_basic(const std::string& name) {
+  if (m_types.find(name) == m_types.end()) {
+    m_forward_declared_types[name] = BASIC;
+  }
+}
+
+/*!
+ * Inform the type system that there will eventually be a type named "name" and that it's a
+ * structure. This allows the type to be used in a few specific places. For instance a structure can
+ * have a field where an element is the same type.
+ */
+void TypeSystem::forward_declare_type_as_structure(const std::string& name) {
+  if (m_types.find(name) == m_types.end()) {
+    m_forward_declared_types[name] = STRUCTURE;
+  }
 }
 
 /*!
@@ -121,7 +145,7 @@ DerefInfo TypeSystem::get_deref_info(const TypeSpec& ts) {
   } else if (ts.base_type() == "pointer") {
     info.can_deref = true;
     info.result_type = ts.get_single_arg();
-    auto result_type = lookup_type(info.result_type);
+    auto result_type = lookup_type_allow_partial_def(info.result_type);
     if (result_type->is_reference()) {
       // in memory, an array of pointers
       info.stride = POINTER_SIZE;
@@ -158,6 +182,10 @@ TypeSpec TypeSystem::make_typespec(const std::string& name) const {
 
 bool TypeSystem::fully_defined_type_exists(const std::string& name) const {
   return m_types.find(name) != m_types.end();
+}
+
+bool TypeSystem::partially_defined_type_exists(const std::string& name) const {
+  return m_forward_declared_types.find(name) != m_forward_declared_types.end();
 }
 
 /*!
@@ -231,6 +259,42 @@ Type* TypeSystem::lookup_type(const std::string& name) const {
  */
 Type* TypeSystem::lookup_type(const TypeSpec& ts) const {
   return lookup_type(ts.base_type());
+}
+
+/*!
+ * Get type info. If the type is not fully defined (ie, we are parsing its deftype now) and its
+ * forward defined as a basic or structure, just get basic/structure.
+ */
+Type* TypeSystem::lookup_type_allow_partial_def(const TypeSpec& ts) const {
+  return lookup_type_allow_partial_def(ts.base_type());
+}
+
+/*!
+ * Get type info. If the type is not fully defined (ie, we are parsing its deftype now) and its
+ * forward defined as a basic or structure, just get basic/structure.
+ */
+Type* TypeSystem::lookup_type_allow_partial_def(const std::string& name) const {
+  // look up fully defined types first:
+  auto kv = m_types.find(name);
+  if (kv != m_types.end()) {
+    return kv->second.get();
+  }
+
+  auto fwd_dec = m_forward_declared_types.find(name);
+  if (fwd_dec != m_forward_declared_types.end()) {
+    if (fwd_dec->second == STRUCTURE) {
+      return lookup_type("structure");
+    } else if (fwd_dec->second == BASIC) {
+      return lookup_type("basic");
+    } else {
+      fmt::print("[TypeSystem] The type {} is not fully define (allow partial).\n", name);
+    }
+
+  } else {
+    fmt::print("[TypeSystem] The type {} is not defined.\n", name);
+  }
+
+  throw std::runtime_error("lookup_type_allow_partial_def failed");
 }
 
 MethodInfo TypeSystem::add_method(const std::string& type_name,
@@ -412,7 +476,7 @@ FieldLookupInfo TypeSystem::lookup_field_info(const std::string& type_name,
     info.array_size = info.field.array_size();
   }
 
-  auto base_type = lookup_type(info.field.type());
+  auto base_type = lookup_type_allow_partial_def(info.field.type());
   if (base_type->is_reference()) {
     if (info.field.is_inline()) {
       if (info.field.is_array()) {
@@ -542,11 +606,7 @@ void TypeSystem::add_builtin_types() {
   auto kheap_type = add_builtin_structure("structure", "kheap");
   auto array_type = add_builtin_basic("basic", "array");
   auto pair_type = add_builtin_structure("object", "pair", true);
-  auto process_tree_type = add_builtin_basic("basic", "process-tree");
-  auto process_type = add_builtin_basic("process-tree", "process");
-  auto thread_type = add_builtin_basic("basic", "thread");
   auto connectable_type = add_builtin_structure("structure", "connectable");
-  auto stack_frame_type = add_builtin_basic("basic", "stack-frame");
   auto file_stream_type = add_builtin_basic("basic", "file-stream");
   add_builtin_value_type("object", "pointer", 4);
   auto inline_array_type = add_builtin_value_type("object", "inline-array", 4);
@@ -566,7 +626,7 @@ void TypeSystem::add_builtin_types() {
   add_builtin_value_type("uinteger", "uint8", 1);
   add_builtin_value_type("uinteger", "uint16", 2);
   add_builtin_value_type("uinteger", "uint32", 4);
-  add_builtin_value_type("uinteger", "uint64", 81);
+  add_builtin_value_type("uinteger", "uint64", 8);
   add_builtin_value_type("uinteger", "uint128", 16, false, false, RegKind::INT_128);
 
   auto int_type = add_builtin_value_type("integer", "int", 8, false, true);
@@ -585,7 +645,7 @@ void TypeSystem::add_builtin_types() {
              make_function_typespec({"_type_"}, "int"));  // todo - this integer type?
   add_method(obj_type, "asize-of", make_function_typespec({"_type_"}, "int"));
   add_method(obj_type, "copy", make_function_typespec({"_type_", "symbol"}, "_type_"));
-  add_method(obj_type, "relocate", make_function_typespec({"_type_", "int32"}, "_type_"));
+  add_method(obj_type, "relocate", make_function_typespec({"_type_", "int"}, "_type_"));
   add_method(obj_type, "mem-usage",
              make_function_typespec({"_type_"}, "int32"));  // todo - this is a guess.
 
@@ -660,11 +720,7 @@ void TypeSystem::add_builtin_types() {
   add_field_to_type(pair_type, "cdr", make_typespec("object"));
 
   // todo, with kernel
-  (void)process_tree_type;
-  (void)process_type;
-  (void)thread_type;
   (void)connectable_type;
-  (void)stack_frame_type;
   (void)file_stream_type;
 }
 
@@ -717,7 +773,7 @@ Field TypeSystem::lookup_field(const std::string& type_name, const std::string& 
  * Get the minimum required aligment of a field.
  */
 int TypeSystem::get_alignment_in_type(const Field& field) {
-  auto field_type = lookup_type(field.type());
+  auto field_type = lookup_type_allow_partial_def(field.type());
 
   if (field.is_inline()) {
     if (field.is_array()) {
@@ -727,7 +783,7 @@ int TypeSystem::get_alignment_in_type(const Field& field) {
     } else {
       // it is an inlined field, so return the alignment in memory
       // TODO - for inline, but not inline array, do we use structure alignment always?
-      return field_type->get_in_memory_alignment();
+      return field_type->get_inline_array_alignment();
     }
   }
 
@@ -740,18 +796,31 @@ int TypeSystem::get_alignment_in_type(const Field& field) {
   return POINTER_SIZE;
 }
 
+namespace {
+bool allow_inline(const Type* type) {
+  auto name = type->get_name();
+  return name != "basic" && name != "structure";
+}
+}  // namespace
+
 /*!
  * Get the size of a field in a type.  The array sizes should be consistent with get_deref_info's
  * stride.
  */
-int TypeSystem::get_size_in_type(const Field& field) {
+int TypeSystem::get_size_in_type(const Field& field) const {
   if (field.is_dynamic()) {
     return 0;
   }
-  auto field_type = lookup_type(field.type());
+  auto field_type = lookup_type_allow_partial_def(field.type());
 
   if (field.is_array()) {
     if (field.is_inline()) {
+      if (!allow_inline(field_type)) {
+        fmt::print(
+            "[Type System] Attempted to use {} inline, this probably isn't what you wanted.\n",
+            field_type->get_name());
+        throw std::runtime_error("bad get size in type");
+      }
       assert(field_type->is_reference());
       return field.array_size() *
              align(field_type->get_size_in_memory(), field_type->get_inline_array_alignment());
@@ -766,8 +835,16 @@ int TypeSystem::get_size_in_type(const Field& field) {
   } else {
     // not an array
     if (field.is_inline()) {
+      if (!allow_inline(field_type)) {
+        fmt::print(
+            "[Type System] Attempted to use {} inline, this probably isn't what you wanted.\n",
+            field_type->get_name());
+        throw std::runtime_error("bad get size in type");
+      }
       assert(field_type->is_reference());
-      return align(field_type->get_size_in_memory(), field_type->get_in_memory_alignment());
+      // return align(field_type->get_size_in_memory(), field_type->get_in_memory_alignment());
+      // looking at dead-pool-heap we tightly pack in this case
+      return field_type->get_size_in_memory();
     } else {
       if (field_type->is_reference()) {
         return POINTER_SIZE;
@@ -888,16 +965,16 @@ bool TypeSystem::typecheck_base_types(const std::string& expected,
   // declared, but not defined?)
   lookup_type(expected);
 
-  if (expected == actual) {
-    lookup_type(actual);  // make sure it exists
+  if (expected == actual || expected == lookup_type_allow_partial_def(actual)->get_name()) {
+    lookup_type_allow_partial_def(actual);  // make sure it exists
     return true;
   }
 
   std::string actual_name = actual;
-  auto actual_type = lookup_type(actual_name);
+  auto actual_type = lookup_type_allow_partial_def(actual_name);
   while (actual_type->has_parent()) {
     actual_name = actual_type->get_parent();
-    actual_type = lookup_type(actual_name);
+    actual_type = lookup_type_allow_partial_def(actual_name);
 
     if (expected == actual_name) {
       return true;
@@ -991,12 +1068,12 @@ TypeSpec TypeSystem::lowest_common_ancestor(const std::vector<TypeSpec>& types) 
 TypeSpec coerce_to_reg_type(const TypeSpec& in) {
   if (in.arg_count() == 0) {
     if (in.base_type() == "int8" || in.base_type() == "int16" || in.base_type() == "int32" ||
-        in.base_type() == "int16") {
+        in.base_type() == "int64") {
       return TypeSpec("int");
     }
 
     if (in.base_type() == "uint8" || in.base_type() == "uint16" || in.base_type() == "uint32" ||
-        in.base_type() == "uint16") {
+        in.base_type() == "uint64") {
       return TypeSpec("uint");
     }
   }