+++ common/type_system/deftype.h
@@ -19,6 +19,7 @@ struct DeftypeResult {
   TypeFlags flags;
   TypeSpec type;
   Type* type_info = nullptr;
+  bool create_runtime_type = true;
 };
 
 DeftypeResult parse_deftype(const goos::Object& deftype, TypeSystem* ts);